add v4.3.0 logic

+++ iOS/AgoraEntScenarios/Scenes/KTV/View/KTV/View/VLVoicePerShowView.h
@@ -19,7 +19,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)didVolQualityGradeChangedWithIndex:(NSInteger)index;
 - (void)voiceDelaySelectedAction:(BOOL)isSelected;
 -(void)didAECStateChange:(BOOL)enable;
--(void)didAECLevelSetWith:(int)level;
+-(void)didAECLevelSetWith:(NSInteger)level;
 @end
 
 @interface VLVoicePerShowView : UIView
